Fix: remove @obj class and extension signatures (#1448)
* WIP rm obj

* convert caller

* converted ZMClientRegistrationStatus

* converted testThatItReturns_FetchingClients_WhenReceivingAnErrorWithTooManyClients

* clean up

* fix test target not found UserChangeInfo.addObserver method issue

* clean up

* Update ConversationListChangeInfo+UserSession.swift

* Update MessageChangeInfo+UserSession.swift

* refactoring

* fix test

* Update UserClientRequestStrategy.swift

* Update PreLoginAuthenticationNotification.swift

* Update UserChangeInfo+UserSession.swift
