+++ Source/Data Model/MessageChangeInfo+UserSession.swift
@@ -1,6 +1,6 @@
 //
 // Wire
-// Copyright (C) 2017 Wire Swiss GmbH
+// Copyright (C) 2021 Wire Swiss GmbH
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -24,7 +24,6 @@ extension MessageChangeInfo {
     /// To observe messages and their users (senders, systemMessage users), observe the conversation window instead
     /// Messages observed with this call will not contain information about user changes
     /// You must hold on to the token and use it to unregister
-    @objc(addObserver:forMessage:userSession:)
     public static func add(observer: ZMMessageObserver,
                            for message: ZMConversationMessage,
                            userSession: ZMUserSession) -> NSObjectProtocol {
@@ -36,7 +35,6 @@ extension NewUnreadMessagesChangeInfo {
     
     /// Adds a ZMNewUnreadMessagesObserver
     /// You must hold on to the token and use it to unregister
-    @objc(addNewMessageObserver:forUserSession:)
     public static func add(observer: ZMNewUnreadMessagesObserver, for userSession: ZMUserSession) -> NSObjectProtocol {
         return self.add(observer: observer, managedObjectContext: userSession.managedObjectContext)
     }
@@ -46,7 +44,6 @@ extension NewUnreadMessagesChangeInfo {
 extension NewUnreadKnockMessagesChangeInfo {
     /// Adds a ZMNewUnreadKnocksObserver
     /// You must hold on to the token and use it to unregister
-    @objc(addNewKnockObserver:forUserSession:)
     public static func add(observer: ZMNewUnreadKnocksObserver, for userSession: ZMUserSession) -> NSObjectProtocol {
         return self.add(observer: observer, managedObjectContext: userSession.managedObjectContext)
     }
