+++ Source/Data Model/UserChangeInfo+UserSession.swift
@@ -1,6 +1,6 @@
 //
 // Wire
-// Copyright (C) 2017 Wire Swiss GmbH
+// Copyright (C) 2021 Wire Swiss GmbH
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -26,7 +26,6 @@ extension UserChangeInfo {
     /// Adds an observer for a user conforming to UserType. You must hold on to the token until you want to stop
     /// observing.
     ///
-    @objc(addObserver:forUser:inUserSession:)
     public static func add(observer: ZMUserObserver, for user: UserType, in userSession: ZMUserSession) -> NSObjectProtocol? {
         return add(observer: observer, for: user, in: userSession.managedObjectContext)
     }
@@ -35,7 +34,6 @@ extension UserChangeInfo {
 
     /// Adds an observer for changes in all ZMUsers. You must hold on to the token until you want to stop observing.
     ///
-    @objc(addUserObserver:inUserSession:)
     public static func add(userObserver observer: ZMUserObserver, in userSession: ZMUserSession) -> NSObjectProtocol {
         return add(userObserver: observer, in: userSession.managedObjectContext)
     }
@@ -45,7 +43,6 @@ extension UserChangeInfo {
     /// Adds an observer for changes in all ZMSearchUsers. You must hold on to the token until you want to stop
     /// observing.
     ///
-    @objc(addSearchUserObserver:inUserSession:)
     public static func add(searchUserObserver observer: ZMUserObserver, in userSession: ZMUserSession) -> NSObjectProtocol {
         return add(searchUserObserver: observer, in: userSession.managedObjectContext)
     }
