+++ Source/Synchronization/Strategies/DeleteAccountRequestStrategy.swift
@@ -21,7 +21,7 @@ import Foundation
 import WireTransport
 
 /// Requests the account deletion
-@objc public final class DeleteAccountRequestStrategy: AbstractRequestStrategy, ZMSingleRequestTranscoder {
+public final class DeleteAccountRequestStrategy: AbstractRequestStrategy, ZMSingleRequestTranscoder {
 
     fileprivate static let path: String = "/self"
     public static let userDeletionInitiatedKey: String = "ZMUserDeletionInitiatedKey"
