+++ Source/Synchronization/Strategies/FeatureFlagRequestStrategy.swift
@@ -19,8 +19,7 @@
 import Foundation
 
 // Get FeatureFlags
-@objc
-public final class FeatureFlagRequestStrategy: AbstractRequestStrategy {
+public final class FeatureFlagRequestStrategy: AbstractRequestStrategy, ZMSingleRequestTranscoder {
     
     // MARK: - Private Property
     private let syncContext: NSManagedObjectContext
@@ -66,10 +65,9 @@ public final class FeatureFlagRequestStrategy: AbstractRequestStrategy {
         singleRequestSync.readyForNextRequestIfNotBusy()
         return singleRequestSync.nextRequest()
     }
-}
 
-// MARK: - ZMSingleRequestTranscoder
-extension FeatureFlagRequestStrategy: ZMSingleRequestTranscoder {
+    // MARK: - ZMSingleRequestTranscoder
+    
     public func request(for sync: ZMSingleRequestSync) -> ZMTransportRequest? {
         switch sync {
         case singleRequestSync:
