+++ Source/Synchronization/Strategies/LabelUpstreamRequestStrategy.swift
@@ -18,8 +18,7 @@
 
 import Foundation
 
-@objc
-public class LabelUpstreamRequestStrategy: AbstractRequestStrategy {
+public class LabelUpstreamRequestStrategy: AbstractRequestStrategy, ZMContextChangeTracker, ZMContextChangeTrackerSource, ZMSingleRequestTranscoder {
     
     fileprivate let jsonEncoder = JSONEncoder()
     fileprivate var upstreamSync: ZMSingleRequestSync!
@@ -35,9 +34,7 @@ public class LabelUpstreamRequestStrategy: AbstractRequestStrategy {
         return upstreamSync.nextRequest()
     }
     
-}
-
-extension LabelUpstreamRequestStrategy: ZMContextChangeTracker, ZMContextChangeTrackerSource {
+    //MARK:- ZMContextChangeTracker, ZMContextChangeTrackerSource
     
     public var contextChangeTrackers: [ZMContextChangeTracker] {
         return [self]
@@ -65,9 +62,7 @@ extension LabelUpstreamRequestStrategy: ZMContextChangeTracker, ZMContextChangeT
         upstreamSync.readyForNextRequestIfNotBusy()
     }
     
-}
-
-extension LabelUpstreamRequestStrategy: ZMSingleRequestTranscoder {
+//MARK:- ZMSingleRequestTranscoder
     
     public func request(for sync: ZMSingleRequestSync) -> ZMTransportRequest? {
         let fetchRequest = NSFetchRequest<Label>(entityName: Label.entityName())
