+++ Source/Synchronization/Strategies/LegalHoldRequestStrategy.swift
@@ -18,8 +18,7 @@
 
 import Foundation
 
-@objc
-public class LegalHoldRequestStrategy: AbstractRequestStrategy {
+public class LegalHoldRequestStrategy: AbstractRequestStrategy, ZMSingleRequestTranscoder, ZMEventConsumer {
     
     fileprivate let syncStatus: SyncStatus
     fileprivate var singleRequstSync: ZMSingleRequestSync!
@@ -41,11 +40,7 @@ public class LegalHoldRequestStrategy: AbstractRequestStrategy {
         
         return singleRequstSync.nextRequest()
     }
-    
-}
-
-extension LegalHoldRequestStrategy: ZMSingleRequestTranscoder {
-        
+            
     public func request(for sync: ZMSingleRequestSync) -> ZMTransportRequest? {
         let selfUser = ZMUser.selfUser(in: managedObjectContext)
         
@@ -97,9 +92,7 @@ extension LegalHoldRequestStrategy: ZMSingleRequestTranscoder {
         }
     }
     
-}
-
-extension LegalHoldRequestStrategy: ZMEventConsumer {
+    //MARK:- ZMEventConsumer
     
     public func processEvents(_ events: [ZMUpdateEvent], liveEvents: Bool, prefetchResult: ZMFetchRequestBatchResult?) {
         events.forEach(processUpdateEvent)
