+++ Source/Synchronization/Strategies/PermissionsDownloadRequestStrategy.swift
@@ -77,7 +77,7 @@ fileprivate extension Member {
 
 }
 
-public final class PermissionsDownloadRequestStrategy: AbstractRequestStrategy, ZMContextChangeTrackerSource, ZMRequestGeneratorSource {
+public final class PermissionsDownloadRequestStrategy: AbstractRequestStrategy, ZMContextChangeTrackerSource, ZMRequestGeneratorSource, ZMDownstreamTranscoder {
 
     fileprivate var sync: ZMDownstreamObjectSync!
 
@@ -105,9 +105,7 @@ public final class PermissionsDownloadRequestStrategy: AbstractRequestStrategy, 
         return [sync]
     }
 
-}
-
-extension PermissionsDownloadRequestStrategy: ZMDownstreamTranscoder {
+    //MARK:- ZMDownstreamTranscoder
 
     public func request(forFetching object: ZMManagedObject!, downstreamSync: ZMObjectSync!) -> ZMTransportRequest! {
         guard let member = object as? Member, downstreamSync as? ZMDownstreamObjectSync == sync else { fatal("Wrong object: \(object.safeForLoggingDescription)") }
