+++ Source/Synchronization/Strategies/ProxiedRequestStrategy.swift
@@ -37,7 +37,7 @@ extension ProxiedRequestType {
 }
 
 /// Perform requests to the Giphy search API
-@objc public final class ProxiedRequestStrategy : AbstractRequestStrategy {
+public final class ProxiedRequestStrategy : AbstractRequestStrategy {
     
     static fileprivate let BasePath = "/proxy"
