+++ Source/Synchronization/Strategies/SignatureRequestStrategy.swift
@@ -20,8 +20,7 @@
 import Foundation
 
 // Sign a PDF document
-@objc
-public final class SignatureRequestStrategy: AbstractRequestStrategy {
+public final class SignatureRequestStrategy: AbstractRequestStrategy, ZMSingleRequestTranscoder {
     
     // MARK: - Private Property
     private let syncContext: NSManagedObjectContext
@@ -76,10 +75,9 @@ public final class SignatureRequestStrategy: AbstractRequestStrategy {
         }
         return nil
     }
-}
 
-// MARK: - ZMSingleRequestTranscoder
-extension SignatureRequestStrategy: ZMSingleRequestTranscoder {
+    // MARK: - ZMSingleRequestTranscoder
+
     public func request(for sync: ZMSingleRequestSync) -> ZMTransportRequest? {
         switch sync {
         case requestSync:
