+++ Tests/Source/Integration/UserProfileTests.h
@@ -1,6 +1,6 @@
 //
 // Wire
-// Copyright (C) 2019 Wire Swiss GmbH
+// Copyright (C) 2021 Wire Swiss GmbH
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -16,12 +16,11 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-import Foundation
+#import "NSError+ZMUserSession.h"
+#import "NSError+ZMUserSessionInternal.h"
+#import "ZMClientRegistrationStatus+Internal.h"
+#import "ZMCredentials.h"
 
-extension LinkPreviewAssetUploadRequestStrategy {
+@interface UserProfileTests : IntegrationTest
 
-    public static func create(withManagedObjectContext managedObjectContext: NSManagedObjectContext, applicationStatus: ApplicationStatus) -> LinkPreviewAssetUploadRequestStrategy {
-        return LinkPreviewAssetUploadRequestStrategy(managedObjectContext: managedObjectContext, applicationStatus: applicationStatus, linkPreviewPreprocessor: nil, previewImagePreprocessor: nil)
-    }
-
-}
+@end
