+++ Tests/Source/Integration/UserProfileTests.m
@@ -16,17 +16,8 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 // 
 
-#import "NSError+ZMUserSession.h"
-#import "NSError+ZMUserSessionInternal.h"
 #import "Tests-Swift.h"
-#import "ZMClientRegistrationStatus+Internal.h"
-#import "ZMCredentials.h"
-
-
-@interface UserProfileTests : IntegrationTest
-
-@end
-
+#import "UserProfileTests.h"
 
 @implementation UserProfileTests
