+++ Tests/Source/Test-Bridging-Header.h
@@ -27,3 +27,4 @@
 #import "IntegrationTest.h"
 #import "APNSTestsBase.h"
 #import "ZMClientRegistrationStatusTests.h"
+#import "UserProfileTests.h"
